Move remaining parameters (excluding reducers).
* BatchSize
* By
* FL
* Distinct/Group
