+++ src/test/java/com/github/thesench/solr/dsl/stream/expr/decorators/TopTest.java
@@ -2,7 +2,7 @@ package com.github.thesench.solr.dsl.stream.expr.decorators;
 
 import static com.github.thesench.solr.dsl.stream.expr.decorators.StreamDecorators.top;
 import static com.github.thesench.solr.dsl.stream.expr.params.RequestHandler.EXPORT;
-import static com.github.thesench.solr.dsl.stream.expr.params.StreamParameters.fl;
+import static com.github.thesench.solr.dsl.stream.expr.params.FL.fl;
 import static com.github.thesench.solr.dsl.stream.expr.params.N.n;
 import static com.github.thesench.solr.dsl.stream.expr.params.Q.q;
 import static com.github.thesench.solr.dsl.stream.expr.params.QT.qt;
